separated test from twitter.js into twitter_test.js
